+++ Specifications/Kernel/Projections/Engine/Expressions/EventValues/for_EventContextPropertyExpressionResolver/when_asking_can_resolve_for_add_expression.cs
@@ -1,16 +1,16 @@
 // Copyright (c) Aksio Insurtech. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Aksio.Cratis.Events.Projections.Expressions.for_AddExpressionResolver;
+namespace Aksio.Cratis.Events.Projections.Expressions.EventValues.for_EventContextPropertyExpressionResolver;
 
 public class when_asking_can_resolve_for_add_expression : Specification
 {
-    AddExpressionResolver resolver;
+    EventContextPropertyExpressionResolver resolver;
     bool result;
 
     void Establish() => resolver = new();
 
-    void Because() => result = resolver.CanResolve(string.Empty, "$add(something)");
+    void Because() => result = resolver.CanResolve("$eventContext(something)");
 
     [Fact] void should_be_able_to_resolve() => result.ShouldBeTrue();
 }
