+++ Specifications/Kernel/Projections/Engine/Expressions/ModelProperties/for_AddExpressionResolver/when_asking_can_resolve_for_add_expression.cs
@@ -1,0 +1,21 @@
+// Copyright (c) Aksio Insurtech. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace Aksio.Cratis.Events.Projections.Expressions.ModelProperties.for_AddExpressionResolver;
+
+public class when_asking_can_resolve_for_add_expression : Specification
+{
+    Mock<IEventValueProviderExpressionResolvers> event_value_resolvers;
+    AddExpressionResolver resolver;
+    bool result;
+
+    void Establish()
+    {
+        event_value_resolvers = new();
+        resolver = new(event_value_resolvers.Object);
+    }
+
+    void Because() => result = resolver.CanResolve(string.Empty, "$add(something)");
+
+    [Fact] void should_be_able_to_resolve() => result.ShouldBeTrue();
+}
