+++ Specifications/Kernel/Projections/Engine/Expressions/ModelProperties/for_CountExpressionResolver/when_asking_can_resolve_for_count_expression.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Aksio Insurtech. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Aksio.Cratis.Events.Projections.Expressions.for_CountExpressionResolver;
+namespace Aksio.Cratis.Events.Projections.Expressions.ModelProperties.for_CountExpressionResolver;
 
 public class when_asking_can_resolve_for_count_expression : Specification
 {
