+++ Specifications/Kernel/Projections/Engine/Expressions/ModelProperties/for_SubtractExpressionResolver/when_asking_can_resolve_for_add_expression.cs
@@ -1,14 +1,19 @@
 // Copyright (c) Aksio Insurtech. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Aksio.Cratis.Events.Projections.Expressions.for_SubtractExpressionResolver;
+namespace Aksio.Cratis.Events.Projections.Expressions.ModelProperties.for_SubtractExpressionResolver;
 
 public class when_asking_can_resolve_for_add_expression : Specification
 {
+    Mock<IEventValueProviderExpressionResolvers> event_value_resolvers;
     SubtractExpressionResolver resolver;
     bool result;
 
-    void Establish() => resolver = new();
+    void Establish()
+    {
+        event_value_resolvers = new();
+        resolver = new(event_value_resolvers.Object);
+    }
 
     void Because() => result = resolver.CanResolve(string.Empty, "$subtract(something)");
