+++ Specifications/Kernel/Projections/Engine/Expressions/for_ModelPropertyExpressionResolvers/when_asking_can_resolve_for_unknown_expression.cs
@@ -1,14 +1,11 @@
 // Copyright (c) Aksio Insurtech. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Aksio.Cratis.Events.Projections.Expressions.for_PropertyMapperExpressionResolvers;
+namespace Aksio.Cratis.Events.Projections.Expressions.for_ModelPropertyExpressionResolvers;
 
-public class when_asking_can_resolve_for_unknown_expression : Specification
+public class when_asking_can_resolve_for_unknown_expression : given.model_property_expression_resolvers
 {
-    PropertyMapperExpressionResolvers resolvers;
     bool result;
-
-    void Establish() => resolvers = new PropertyMapperExpressionResolvers();
 
     void Because() => result = resolvers.CanResolve(string.Empty, "$randomUnknownExpression");
