+++ Specifications/Kernel/Projections/Engine/Expressions/for_ModelPropertyExpressionResolvers/when_trying_to_resolve_unknown_expression.cs
@@ -1,0 +1,13 @@
+// Copyright (c) Aksio Insurtech. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace Aksio.Cratis.Events.Projections.Expressions.for_ModelPropertyExpressionResolvers;
+
+public class when_trying_to_resolve_unknown_expression : given.model_property_expression_resolvers
+{
+    Exception result;
+
+    void Because() => result = Catch.Exception(() => resolvers.Resolve(string.Empty, "$randomUnknownExpression"));
+
+    [Fact] void should_throw_unsupported_event_value_expression() => result.ShouldBeOfExactType<UnsupportedModelPropertyExpression>();
+}
